requirements.txt
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-cors-headers==3.0.2
psycopg2==2.7.7
psycopg2-binary==2.8.3
pyotp==2.3.0
passlib==1.7.1